+++ packages/workflow-core/src/lib/plugins/common-plugin/transformer-plugin.ts
@@ -27,7 +27,7 @@ export class TransformerPlugin implements ISerializableMappingPluginParams {
     this.stateNames = params.stateNames;
   }
 
-  async invoke(context: TContext) {
+  async invoke(context: TContext, config: unknown) {
     console.log('invoke() method called');
 
     for (const transformParams of this.transformers) {
@@ -36,7 +36,7 @@ export class TransformerPlugin implements ISerializableMappingPluginParams {
         transformParams.mapping,
       );
 
-      transformer.transform(context);
+      transformer.transform({ ...context, workflowRuntimeConfig: config });
     }
 
     console.log('Transform performed successfully.');